+++ sykdomstidslinje/src/main/kotlin/no/nav/helse/sykdomstidslinje/Utbetalingsberegning.kt
@@ -1,6 +1,5 @@
 package no.nav.helse.sykdomstidslinje
 
-import java.math.BigDecimal
 import java.time.LocalDate
 
 data class Utbetalingsberegning(
@@ -11,5 +10,5 @@ data class Utbetalingsberegning(
 data class Utbetalingslinje(
     val fom: LocalDate,
     val tom: LocalDate,
-    val dagsats: BigDecimal
+    val dagsats: Int
 )
